% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-competition-maps.R
\docType{data}
\name{competition_maps}
\alias{competition_maps}
\title{RoboCupRescue Simulation competition maps}
\format{
A \code{data.frame} with 15 rows and 3 columns:
\describe{
\item{year}{An integer representing the year of the competition.}
\item{map}{A character string indicating the name of the map data.}
\item{url}{A character string containing the official download URL for the
map data.}
}
}
\usage{
competition_maps
}
\description{
A dataset containing information about the maps used in past RoboCupRescue
Simulation competitions. It includes the competitions year, the name of
the map, and the official URL to download the map data.
}
\examples{
# Show the dataset
competition_maps

}
\keyword{datasets}
