% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-elements.R
\name{get_elements}
\alias{get_elements}
\alias{get_nodes}
\alias{get_edges}
\alias{get_roads}
\alias{get_buildings}
\alias{get_entrances}
\alias{get_passage}
\title{Get elements from a RoboCupRescue Simulation map}
\usage{
get_nodes(map, ids = character())

get_edges(map, ids = character())

get_roads(map, ids = character())

get_buildings(map, ids = character())

get_entrances(map, ids = character())

get_passage(map, ids = character())
}
\arguments{
\item{map}{An object of class \code{rrs_map}.}

\item{ids}{A character vector of element IDs to filter by.
If empty (default), all elements of the specified type are returned.}
}
\value{
An \code{sf} data frame containing the requested map elements.
}
\description{
These functions provide methods to extract specific geographical elements
from an \code{rrs_map} object. You can retrieve all elements of a certain type
or filter them by their IDs.
}
\examples{
gml <- system.file("extdata", "map-test.gml", package = "rrstools")
map <- read_rrs_map(gml)

# Get all nodes
get_nodes(map)

# Get a specific node by ID
get_nodes(map, c("0", "2"))
}
