% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub-url.R
\name{get_hub_url}
\alias{get_hub_url}
\title{Get the URL of the RCRS Scenario Hub}
\usage{
get_hub_url()
}
\value{
A string of the current URL of the RCRS Scenario Hub.
}
\description{
\code{get_hub_url} returns the current URL of the RCRS Scenario Hub.
}
\examples{
get_hub_url()
}
