% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morph-road-network.R
\name{morph_road_network}
\alias{morph_road_network}
\title{Morph raw map data into a road network}
\usage{
morph_road_network(map_data)
}
\arguments{
\item{map_data}{\code{rrs_map} data, the GML map data loaded by
the \code{read_rrs_map()}.}
}
\value{
A list containing the network components:
\itemize{
\item \code{nodes}: A sf object of nodes with point geometries.
\item \code{edges}: A sf object of edges with linestring geometries.
}
}
\description{
\code{morph_road_network()} converts map data from the RoboCup Rescue Simulation
into a road network.
}
\details{
This road network consists of nodes and edges. Nodes are defined as
the centroids of all areas and the center points the boundaries
between areas. Edges connect nodes between an agent can travel.
}
\examples{
gml <- system.file("extdata", "map-test.gml", package = "rrstools")
map_data <- read_rrs_map(gml)
network <- morph_road_network(map_data)
network
}
