% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrs-scenario.R
\name{plot.rrs_scenario}
\alias{plot.rrs_scenario}
\title{Plot a RRS scenario object (Not supported directly)}
\usage{
# Collect usage:
# plot(map, scenario)
}
\arguments{
\item{x}{An object class \code{rrs_scenario}.}

\item{...}{Not used.}
}
\description{
A \code{rrs_scenario} object cannot be plotted by itself because it lacks the
necessary spatial conext provided by a map. This function exists to
intercept calls to \code{plot(scenario)} and provide a helpful error message.
}
\details{
To visualize a scenario, you must plot it together with a \code{rrs_map}
object.
}
