% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrs-map.R
\name{print.rrs_map}
\alias{print.rrs_map}
\title{Print a RoboCupRescue Simulation map object}
\usage{
\method{print}{rrs_map}(x, ...)
}
\arguments{
\item{x}{An object class \code{rrs_map}.}

\item{...}{Not used.}
}
\description{
\code{print} prints a concise summary of a \code{rrs_map} object.
It displays the object type, the number of nodes, edges, and faces it
contains, its overall bounding box.
}
\examples{
gml <- system.file("extdata", "map-test.gml", package = "rrstools")
map <- read_rrs_map(gml)
map
}
