% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrs-scenario.R
\name{print.rrs_scenario}
\alias{print.rrs_scenario}
\title{Print a RRS scenario object}
\usage{
\method{print}{rrs_scenario}(x, ...)
}
\arguments{
\item{x}{An object class \code{rrs_scenario}.}

\item{...}{Not used.}
}
\description{
\code{print} prints a concise summary of a \code{rrs_scenario} object.
It displays the object type, the total number of entities, and a
table with the counts of entities for each type.
}
\examples{
xml <- system.file("extdata", "scenario-test.xml", package = "rrstools")
scenario <- read_rrs_scenario(xml)
scenario
}
