% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rrs_map.R
\name{read_rrs_map}
\alias{read_rrs_map}
\title{Read RoboCupRescue Simulation map data from GML file}
\usage{
read_rrs_map(gml)
}
\arguments{
\item{gml}{Path to the GML file.}
}
\value{
A list of \code{sf} objects: nodes, edges, buildings, and roads.
}
\description{
This function reads and processes map data for RoboCupRescue Simulation
from a GML file. It extracts nodes, edges, buildings and roads, and
organizes them into a list.
}
\examples{
gml <- system.file("extdata", "map.gml", package = "rrstools")
map_data <- read_rrs_map(gml)
map_data
}
