% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-rrs-scenario.R
\name{read_rrs_scenario}
\alias{read_rrs_scenario}
\title{Read RoboCupRescue Simulation scenario data from XML file}
\usage{
read_rrs_scenario(xml)
}
\arguments{
\item{xml}{Path to the XML file containing the scenario data.}
}
\value{
A \code{rrs_scenario} object. This object is a \code{data.frame}, and contains
the following columns:
\itemize{
\item \code{type}: Character. The type of the entity
(e.g., "refuge", "civilian").
\item \code{location}: Character. The entity ID of the location where
the entity is placed.
\item \code{attributes}: List. A list containing other attributes of the
entity.
}
}
\description{
\code{read_rrs_scenario()} reads and processes scenario data for
RoboCupRescue Simulation from a XML file.
}
\examples{
xml <- system.file("extdata", "scenario-test.xml", package = "rrstools")
scenario <- read_rrs_scenario(xml)
scenario
}
