% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrstools-package.R
\docType{package}
\name{rrstools-package}
\alias{rrstools}
\alias{rrstools-package}
\title{rrstools: Analyzing RoboCupRescue Simulation Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Tools for analyzing data from RoboCupRescue Simulation (RRS) \url{https://rescuesim.robocup.org}, a disaster rescue simulation platform. Supports reading virtual city map and disaster scenario files into analyzable data structures and provides functions for their visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nononoexe.github.io/rrstools/}
}

}
\author{
\strong{Maintainer}: Keisuke Ando \email{nononoexe@ymail.ne.jp} (\href{https://orcid.org/0009-0002-4085-2067}{ORCID})

Authors:
\itemize{
  \item Keisuke Ando \email{nononoexe@ymail.ne.jp} (\href{https://orcid.org/0009-0002-4085-2067}{ORCID})
}

}
\keyword{internal}
