% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrs-map.R
\name{plot.rrs_map}
\alias{plot.rrs_map}
\title{Plot a rrs map}
\usage{
\method{plot}{rrs_map}(
  x,
  building_colour = "#f0e7d8",
  building_border = "#121212",
  road_colour = "#dbdbdb",
  road_border = "#121212",
  background_colour = NA,
  ...
)
}
\arguments{
\item{x}{An object class \code{rrs_map}.}

\item{building_colour}{The colour of the buildings.}

\item{building_border}{The border colour of the buildings.}

\item{road_colour}{The colour of the roads.}

\item{road_border}{The border colour of the roads.}

\item{background_colour}{The background colour of the plot.}

\item{...}{Additional arguments passed to \code{par}.}
}
\description{
\code{plot.rrs_map} visualizes a \code{rrs_map} object, displaying the geographical
features such as buildings and roads. It leverages the \code{sf} package's
plotting capbilities to render the map components.
}
\examples{
gml <- system.file("extdata", "map-test.gml", package = "rrstools")
map <- read_rrs_map(gml)
plot(map)
}
