% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-rrs-map.R
\name{read_rrs_map}
\alias{read_rrs_map}
\title{Read RoboCupRescue Simulation map data from GML file}
\usage{
read_rrs_map(gml, scale_data = FALSE)
}
\arguments{
\item{gml}{Path to the GML file.}

\item{scale_data}{Logical. If \code{TRUE}, coordinates are scaled up by a factor
of 1000 to match the simulation environment, and adjusted such that
the minimum x and y values are 0. (Default: \code{TRUE})}
}
\value{
A list of \code{sf} objects: nodes, edges, buildings, and roads.
Also includes a sub-list \code{raw_parsed_data} containing the intermediate
data frames from XML parsing.
}
\description{
This function reads and processes map data for RoboCupRescue Simulation
from a GML file. It extracts nodes, edges, buildings and roads, and
organizes them into a list.
}
\examples{
gml <- system.file("extdata", "map-test.gml", package = "rrstools")
map_data <- read_rrs_map(gml)
map_data
}
