% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrstools-package.R
\docType{package}
\name{rrstools-package}
\alias{rrstools}
\alias{rrstools-package}
\title{rrstools: Tools for Analyzing RoboCupRescue Simulation Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions for analyzing RoboCupRescue Simulation data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nononoexe.github.io/rrstools/}
}

}
\author{
\strong{Maintainer}: Keisuke ANDO \email{ando@maslab.aitech.ac.jp} (\href{https://orcid.org/0009-0002-4085-2067}{ORCID})

}
\keyword{internal}
